requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
exceptiongroup==1.2.2
fastapi==0.112.2
frozenlist==1.4.1
h11==0.14.0
idna==3.8
multidict==6.0.5
paramiko==3.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyNaCl==1.5.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
yarl==1.9.6
python-dateutil
httpx
psutil